refactor: v2 - ai assistant - split tools bridges
